Move pep425tags tests to their own file
